chore: restructure files in 360 images package
