+++ viewer/packages/360-images/src/icons/Image360Icon.ts
