+++ viewer/packages/360-images/src/icons/image360Icon.vert
